Turned on -Wall option when compiling in DEBUG mode

+++ src/projects/base/ovlibrary/data.cpp
@@ -131,11 +131,11 @@ namespace ov
 	{
 		if (offset >= 0)
 		{
-			OV_ASSERT2(GetLength() >= offset);
+			OV_ASSERT2(GetLength() >= static_cast<size_t>(offset));
 		}
 		else
 		{
-			OV_ASSERT2(GetLength() > (-1LL * offset));
+			OV_ASSERT2(GetLength() > static_cast<size_t>(-1LL * offset));
 		}
 
 		return Subdata(offset, (offset >= 0) ? (GetLength() - offset) : (static_cast<size_t>(-1 * offset)));
